Centralize horse-item conversion in BetterHorsesAPI
