+++ BetterHorses/src/main/java/me/luisgamedev/betterhorses/listeners/RightClickListener.java
@@ -1,23 +1,15 @@
 package me.luisgamedev.betterhorses.listeners;
 
 import me.luisgamedev.betterhorses.BetterHorses;
+import me.luisgamedev.betterhorses.api.BetterHorseKeys;
 import me.luisgamedev.betterhorses.api.BetterHorsesAPI;
 import me.luisgamedev.betterhorses.api.events.BetterHorseSpawnEvent;
 import me.luisgamedev.betterhorses.language.LanguageManager;
 import me.luisgamedev.betterhorses.training.TrainingManager;
-import me.luisgamedev.betterhorses.api.BetterHorseKeys;
-import me.luisgamedev.betterhorses.utils.AttributeResolver;
-import me.luisgamedev.betterhorses.utils.HorseArmorUtils;
-import me.luisgamedev.betterhorses.utils.MountConfig;
 import me.luisgamedev.betterhorses.utils.SupportedMountType;
 import org.bukkit.Material;
-import org.bukkit.NamespacedKey;
-import org.bukkit.attribute.Attribute;
-import org.bukkit.attribute.AttributeInstance;
 import org.bukkit.configuration.file.FileConfiguration;
-import org.bukkit.entity.AnimalTamer;
 import org.bukkit.entity.AbstractHorse;
-import org.bukkit.entity.Horse;
 import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
@@ -26,7 +18,6 @@ import org.bukkit.event.player.PlayerInteractEvent;
 import org.bukkit.inventory.EquipmentSlot;
 import org.bukkit.inventory.ItemStack;
 import org.bukkit.inventory.meta.ItemMeta;
-import org.bukkit.persistence.PersistentDataContainer;
 import org.bukkit.persistence.PersistentDataType;
 
 public class RightClickListener implements Listener {
@@ -55,148 +46,32 @@ public class RightClickListener implements Listener {
         if (!item.hasItemMeta() || item.getType() != expectedMaterial) return;
 
         ItemMeta meta = item.getItemMeta();
-        PersistentDataContainer data = meta.getPersistentDataContainer();
-        TrainingManager.ensureTrainingData(data);
+        TrainingManager.ensureTrainingData(meta.getPersistentDataContainer());
         item.setItemMeta(meta);
 
-        Double health = data.get(new NamespacedKey(BetterHorses.getInstance(), "health"), PersistentDataType.DOUBLE);
-        Double currentHealth = data.get(new NamespacedKey(BetterHorses.getInstance(), "current_health"), PersistentDataType.DOUBLE);
-        Double speed = data.get(new NamespacedKey(BetterHorses.getInstance(), "speed"), PersistentDataType.DOUBLE);
-        Double jump = data.get(new NamespacedKey(BetterHorses.getInstance(), "jump"), PersistentDataType.DOUBLE);
-        Double baseHealth = data.get(BetterHorseKeys.BASE_HEALTH, PersistentDataType.DOUBLE);
-        Double baseSpeed = data.get(BetterHorseKeys.BASE_SPEED, PersistentDataType.DOUBLE);
-        Double baseJump = data.get(BetterHorseKeys.BASE_JUMP, PersistentDataType.DOUBLE);
-        Double ridingUnits = data.get(BetterHorseKeys.TRAINING_RIDING_UNITS, PersistentDataType.DOUBLE);
-        Double brushingUnits = data.get(BetterHorseKeys.TRAINING_BRUSHING_UNITS, PersistentDataType.DOUBLE);
-        Double feedingUnits = data.get(BetterHorseKeys.TRAINING_FEEDING_UNITS, PersistentDataType.DOUBLE);
-        long brushTrainingCooldown = data.getOrDefault(BetterHorseKeys.TRAINING_BRUSH_COOLDOWN, PersistentDataType.LONG, 0L);
-        long feedTrainingCooldown = data.getOrDefault(BetterHorseKeys.TRAINING_FEED_COOLDOWN, PersistentDataType.LONG, 0L);
-        String gender = data.get(new NamespacedKey(BetterHorses.getInstance(), "gender"), PersistentDataType.STRING);
-        String ownerUUID = player.getUniqueId().toString();
-        String styleStr = data.get(new NamespacedKey(BetterHorses.getInstance(), "style"), PersistentDataType.STRING);
-        String colorStr = data.get(new NamespacedKey(BetterHorses.getInstance(), "color"), PersistentDataType.STRING);
-        String saddleStr = data.get(new NamespacedKey(BetterHorses.getInstance(), "saddle"), PersistentDataType.STRING);
-        String armorStr = data.get(new NamespacedKey(BetterHorses.getInstance(), "armor"), PersistentDataType.STRING);
-        String customName = data.get(new NamespacedKey(BetterHorses.getInstance(), "name"), PersistentDataType.STRING);
-        String trait = data.get(new NamespacedKey(BetterHorses.getInstance(), "trait"), PersistentDataType.STRING);
-        Byte neutered = data.get(new NamespacedKey(BetterHorses.getInstance(), "neutered"), PersistentDataType.BYTE);
-        int growthStage = data.getOrDefault(new NamespacedKey(BetterHorses.getInstance(), "growth_stage"), PersistentDataType.INTEGER, 10);
-        String mountTypeName = data.get(new NamespacedKey(BetterHorses.getInstance(), "mount_type"), PersistentDataType.STRING);
-        Long cooldown = data.has(new NamespacedKey(BetterHorses.getInstance(), "cooldown"), PersistentDataType.LONG)
-                ? data.get(new NamespacedKey(BetterHorses.getInstance(), "cooldown"), PersistentDataType.LONG)
-                : null;
-
+        Double health = meta.getPersistentDataContainer().get(BetterHorseKeys.HEALTH, PersistentDataType.DOUBLE);
+        Double speed = meta.getPersistentDataContainer().get(BetterHorseKeys.SPEED, PersistentDataType.DOUBLE);
+        Double jump = meta.getPersistentDataContainer().get(BetterHorseKeys.JUMP, PersistentDataType.DOUBLE);
+        String gender = meta.getPersistentDataContainer().get(BetterHorseKeys.GENDER, PersistentDataType.STRING);
+        String mountTypeName = meta.getPersistentDataContainer().get(BetterHorseKeys.MOUNT_TYPE, PersistentDataType.STRING);
         SupportedMountType mountType = SupportedMountType.fromNameOrDefault(mountTypeName);
         String mountName = mountType.getDisplayName(lang);
 
-        if (health == null || speed == null || jump == null || gender == null) {
+        if (health == null || speed == null || jump == null || gender == null || !mountType.isEnabled(config)) {
             player.sendMessage(lang.getFormatted("messages.invalid-horse-data", "%mount%", mountName));
             return;
         }
 
-        if (!mountType.isEnabled(config)) {
-            player.sendMessage(lang.getFormatted("messages.invalid-horse-data", "%mount%", mountName));
-            return;
-        }
-
-        AbstractHorse horse;
-        try {
-            horse = mountType.spawn(player.getLocation());
-        } catch (Exception e) {
-            player.sendMessage(lang.get("messages.cant-spawn"));
-            return;
-        }
-
-        if (horse == null || !horse.isValid()) {
+        AbstractHorse horse = BetterHorsesAPI.toHorse(item, player);
+        if (horse == null) {
             player.sendMessage(lang.get("messages.cant-spawn"));
             return;
         }
 
         BetterHorsesAPI.callSpawnEvent(horse, item.clone(), BetterHorseSpawnEvent.SpawnCause.ITEM);
-
-        double maxScale = config.getDouble("horse-growth-settings.max-size", 1.3);
-        int threshold = config.getInt("horse-growth-settings.ride-and-breed-threshhold", 7);
-        float minScale = (growthStage >= threshold) ? 0.85f : 0.7f;
-        double scale = minScale + ((maxScale - minScale) / 10.0) * growthStage;
-        BetterHorses plugin = BetterHorses.getInstance();
-
-        if (MountConfig.isGrowthEnabled(config, mountType)) {
-            try {
-                setAttribute(horse, Attribute.valueOf("SCALE"), scale);
-            } catch (NoSuchFieldError | Exception e) {
-                plugin.debugLog("RIGHTCLICK_SPAWN", "SET_SCALE", false, e.getMessage());
-            }
-            if (growthStage >= threshold) {
-                horse.setAdult();
-                horse.setAgeLock(false);
-            } else {
-                horse.setBaby();
-                horse.setAgeLock(true);
-            }
-        }
-
-        setAttribute(horse, AttributeResolver.generic("MAX_HEALTH"), health);
-        setAttribute(horse, AttributeResolver.generic("MOVEMENT_SPEED"), speed);
-        setAttribute(horse, Attribute.valueOf("HORSE_JUMP_STRENGTH"), jump);
-        horse.setHealth(currentHealth != null ? currentHealth : health);
-        horse.setTamed(true);
-        horse.setOwner((AnimalTamer) player);
-
-        PersistentDataContainer horseData = horse.getPersistentDataContainer();
-        horseData.set(new NamespacedKey(BetterHorses.getInstance(), "owner"), PersistentDataType.STRING, ownerUUID);
-        horseData.set(new NamespacedKey(BetterHorses.getInstance(), "gender"), PersistentDataType.STRING, gender);
-        horseData.set(new NamespacedKey(BetterHorses.getInstance(), "growth_stage"), PersistentDataType.INTEGER, growthStage == 0 ? 10 : growthStage);
-        horseData.set(new NamespacedKey(BetterHorses.getInstance(), "mount_type"), PersistentDataType.STRING, mountType.getEntityType().name());
-        horseData.set(BetterHorseKeys.BASE_HEALTH, PersistentDataType.DOUBLE, baseHealth != null ? baseHealth : health);
-        horseData.set(BetterHorseKeys.BASE_SPEED, PersistentDataType.DOUBLE, baseSpeed != null ? baseSpeed : speed);
-        horseData.set(BetterHorseKeys.BASE_JUMP, PersistentDataType.DOUBLE, baseJump != null ? baseJump : jump);
-        horseData.set(BetterHorseKeys.TRAINING_RIDING_UNITS, PersistentDataType.DOUBLE, ridingUnits != null ? ridingUnits : 0.0);
-        horseData.set(BetterHorseKeys.TRAINING_BRUSHING_UNITS, PersistentDataType.DOUBLE, brushingUnits != null ? brushingUnits : 0.0);
-        horseData.set(BetterHorseKeys.TRAINING_FEEDING_UNITS, PersistentDataType.DOUBLE, feedingUnits != null ? feedingUnits : 0.0);
-        horseData.set(BetterHorseKeys.TRAINING_BRUSH_COOLDOWN, PersistentDataType.LONG, brushTrainingCooldown);
-        horseData.set(BetterHorseKeys.TRAINING_FEED_COOLDOWN, PersistentDataType.LONG, feedTrainingCooldown);
-
-        if (trait != null && !trait.isBlank()) {
-            horseData.set(new NamespacedKey(BetterHorses.getInstance(), "trait"), PersistentDataType.STRING, trait);
-        }
-
-        if (neutered != null && neutered == (byte) 1) {
-            horseData.set(new NamespacedKey(BetterHorses.getInstance(), "neutered"), PersistentDataType.BYTE, (byte) 1);
-        }
-
-        if (cooldown != null) {
-            horseData.set(new NamespacedKey(BetterHorses.getInstance(), "cooldown"), PersistentDataType.LONG, cooldown);
-        }
-
-        if (customName != null && !customName.isBlank()) {
-            horse.setCustomName(customName);
-            horse.setCustomNameVisible(true);
-        }
-
-        if (horse instanceof Horse h) {
-            try {
-                h.setStyle(Horse.Style.valueOf(styleStr));
-                h.setColor(Horse.Color.valueOf(colorStr));
-            } catch (Exception ignored) {}
-        }
-
-        if (saddleStr != null) {
-            horse.getInventory().setSaddle(new ItemStack(Material.valueOf(saddleStr)));
-        }
-        if (armorStr != null) {
-            HorseArmorUtils.setArmor(horse.getInventory(), new ItemStack(Material.valueOf(armorStr)));
-        }
-
         TrainingManager.recalculateAndApplyBonuses(horse);
 
         item.setAmount(item.getAmount() - 1);
         player.sendMessage(lang.getFormatted("messages.horse-respawned", "%mount%", mountName));
-    }
-
-    private static void setAttribute(AbstractHorse horse, Attribute attribute, double value) {
-        AttributeInstance attr = horse.getAttribute(attribute);
-        if (attr != null) {
-            attr.setBaseValue(value);
-        }
     }
 }
